Allow 90% as msi score for now
The tests that have troubles are the ones that call the share url
without referer and server name.

Verifying that the referer and server name are correct is not so easy in
the tests (this would require us to be able to get a shared result).
I'll add that in the future, but let's already force us to not allow
more mutations than there are now.